+++ src/main/java/projeto/brisa/teste/entity/Ponto.java
@@ -9,6 +9,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import javax.persistence.Table;
 
 import lombok.Data;
 
@@ -19,7 +20,7 @@ public class Ponto implements Serializable{
 	private static final long serialVersionUID = 1L;
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	
 	
 	@ManyToOne
@@ -32,18 +33,19 @@ public class Ponto implements Serializable{
 	
 	private Date dataCriacao;
 	
-	public Ponto(Integer idPonto,Cliente cli, Endereco end) {
+	public Ponto(Long idPonto,Cliente cli, Endereco end) {
 	
 		id = idPonto;
 		cliente = cli;
 		endereco = end;
 		setDataCriacao(new Date());
+
 	}
 	public Ponto() {
 		setDataCriacao(new Date());
 	}
 	
-	public void setIdCliente(Integer id_cliente) {
+	public void setIdCliente(Long id_cliente) {
 		cliente.setId(id_cliente);
 	}
 	public void setIdEndereco(Integer id_endereco) {
